finished backend. no testing done, internet not available
